Acabado

+++ src/aplicacion/Principal.java
@@ -1,10 +1,10 @@
 package aplicacion;
 
-import dominio.Mates;
+import dominio.Matematicas;
 
 public class Principal {
     public static void main(String[] args) {
-        long pasos = Long.parseLong(args[0]);
+        int pasos = Integer.parseInt(args[0]);
         System.out.println("El número PI es " + Matematicas.generarNumeroPiRecursivo(pasos));
     }
 }
